# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "We don't have a public API, so instead of a real host(explosion.team) we specify localhost:3001. To test the following methods, git clone https://github.com/go-park-mail-ru/2019_1_Escapade, enter the root directory and run 'docker-compose up -d'",
        "title": "Escapade Explosion API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/api",
    "paths": {
        "/avatar": {
            "get": {
                "description": "Get user avatar",
                "summary": "Get user avatar",
                "operationId": "GetImage",
                "responses": {
                    "200": {
                        "description": "Avatar found successfully",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "401": {
                        "description": "Required authorization",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "404": {
                        "description": "Avatar not found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    }
                }
            },
            "post": {
                "description": "Create user avatar",
                "summary": "Create user avatar",
                "operationId": "PostImage",
                "responses": {
                    "201": {
                        "description": "Avatar created successfully",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "401": {
                        "description": "Required authorization",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "500": {
                        "description": "Avatar not found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    }
                }
            }
        },
        "/session": {
            "post": {
                "description": "login",
                "summary": "login",
                "operationId": "Login",
                "responses": {
                    "200": {
                        "description": "Get successfully",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.UserPublicInfo"
                        }
                    },
                    "400": {
                        "description": "invalid name or password",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    }
                }
            },
            "delete": {
                "description": "logout",
                "summary": "logout",
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Get successfully",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "401": {
                        "description": "Require authorization",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Result": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.UserPublicInfo": {
            "type": "object",
            "properties": {
                "bestScore": {
                    "type": "string"
                },
                "bestTime": {
                    "type": "string"
                },
                "difficult": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3
                },
                "photo": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        }
    }
}